requirements.txt
streamlit==1.31.0
pandas==2.2.0
sqlalchemy==2.0.27
python-dateutil==2.8.2